Unsafe conversions removed.

+++ src/graphics/vulkan/vk_descriptors.cpp
@@ -91,7 +91,7 @@ VkDescriptorPool DescriptorAllocatorGrowable::get_pool(VkDevice device)
 		//need to create a new pool
 		newPool = create_pool(device, setsPerPool, ratios);
 
-		setsPerPool = setsPerPool*1.5;
+		setsPerPool = static_cast<uint32_t>(setsPerPool*1.5);
 		if (setsPerPool>4092)
 		{
 			setsPerPool = 4092;
@@ -136,7 +136,7 @@ void DescriptorAllocatorGrowable::init(VkDevice device, uint32_t maxSets, std::s
 
 	VkDescriptorPool newPool = create_pool(device, maxSets, poolRatios);
 
-	setsPerPool = maxSets*1.5; //grow it next allocation
+	setsPerPool = static_cast<uint32_t>(maxSets*1.5); //grow it next allocation
 
 	readyPools.push_back(newPool);
 }
